DjangoRestBlog/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycodestyle==2.9.1
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.8.0
